fix: adjusting type

+++ src/PaymentManager/Payment.Application/Handlers/PostPayments/Dto/PaymentDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace Payment.Application.Handlers.PostPayments.Dto;
 
-public sealed record PaymentDto(Guid Payee,Guid Payer,float Amount);
+public sealed record PaymentDto(Guid Payee,Guid Payer,decimal Amount);
